requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
transformers==4.35.0
torch==2.1.0
numpy<2.0.0  # Принудительно используем NumPy 1.x для совместимости
huggingface-hub==0.17.3
sentencepiece==0.1.99
tqdm==4.66.1
accelerate==0.23.0